Moves Legacy CLI into a `legacy` folder and provides a new entrypoint to the beta CLI
- adds a `--legacy (-l)` CLI option that forces the application to run in legacy mode.
- adds a new warning message when the CLI is run in legacy mode.
