+++ lib/legacy/component-folders.ts
