+++ lib/legacy/generate-suggestions.test.ts
