+++ lib/legacy/generate-suggestions.ts
