+++ lib/legacy/http/__mocks__/fetchComponentFolders.ts
