+++ lib/legacy/http/__mocks__/fetchComponents.ts
