+++ lib/legacy/http/__mocks__/fetchVariants.ts
