+++ lib/legacy/http/fetchComponentFolders.ts
