+++ lib/legacy/http/fetchComponents.ts
