+++ lib/legacy/http/fetchVariants.ts
