+++ lib/legacy/importComponents.ts
