+++ lib/legacy/init/token.test.ts
