+++ lib/legacy/init/token.ts
