+++ lib/legacy/output.ts
