+++ lib/legacy/remove-project.ts
