+++ lib/legacy/replace.test.ts
