+++ lib/legacy/replace.ts
