+++ lib/legacy/utils/cleanFileName.ts
