+++ lib/legacy/utils/createSentryContext.ts
