+++ lib/legacy/utils/determineModuleType.test.ts
