+++ lib/legacy/utils/determineModuleType.ts
