+++ lib/legacy/utils/generateJsDriver.ts
