+++ lib/legacy/utils/generateJsDriverTypeFile.ts
