+++ lib/legacy/utils/getSelectedProjects.ts
