+++ lib/legacy/utils/processMetaOption.test.ts
