+++ lib/legacy/utils/processMetaOption.ts
