+++ lib/legacy/utils/projectsToText.ts
