+++ lib/legacy/utils/promptForProject.ts
