+++ lib/legacy/utils/quit.ts
