+++ lib/legacy/utils/sourcesToText.ts
