+++ lib/src/commands/pull.ts
@@ -1,0 +1,3 @@
+export const pull = async () => {
+  console.log("pull");
+};
